cloud/pubsub/example/pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Marcelo A. Amancio <30930605+marceloamancio@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-pubsub = "^2.14.0"
argparse = "^1.4.0"
strenum = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
